int svm input type

+++ model_composition/image_driver_2/containerized/containers/impl/opencv_sift_feats_container.py
@@ -28,7 +28,7 @@ class SIFTFeaturizationContainer(rpc.ModelContainerBase):
         grayscale_img = cv2.cvtColor(img, cv2.COLOR_RGB2GRAY)
         keypoints, features = self.sift.detectAndCompute(grayscale_img, None)
         print(features)
-        return np.array(features[:NUM_SIFT_FEATURES], dtype=np.float32)
+        return np.array(features[:NUM_SIFT_FEATURES], dtype=np.int32)
 
 if __name__ == "__main__":
     print("Starting OpenCV SIFT Featurization Container")
